requirements.txt
# Dependecies
Django>=4.2.8
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
django-redis>=5.4.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
djoser>=2.2.0
Pillow>=10.0.0


# Code quality tools
flake8>=6.1.0
black>=23.11.0